# Databases/static/swagger.yaml
openapi: "3.0.0"
info:
  title: "Event Management API"
  description: "API documentation for the Event Management system."
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:5000"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /events/tags:
    post:
      summary: "Add multiple tags to an event"
      tags:
        - Events
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_id:
                  type: integer
                  example: 1
                tags:
                  type: array
                  items:
                    type: string
                  example: ["urgent", "team"]
      responses:
        '201':
          description: "Tags added successfully"
        '400':
          description: "Missing required fields"
  /events/{event_id}/tags:
    get:
      summary: "Retrieve all tags for a specific event"
      tags:
        - Events
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: "List of tags for the event"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tag_name:
                      type: string
                      example: "urgent"

    put:
      summary: "Update tags for a specific event"
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: event_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
                  example: ["updated", "critical"]
      responses:
        '200':
          description: "Tags updated successfully"
        '400':
          description: "Tags are required"
...
  /messages:
    get:
      summary: "List all messages ordered from newest to oldest"
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: group_id
          required: true
          schema:
            type: integer
            example: 1
          description: "Group ID to filter messages."
        - in: query
          name: page
          required: false
          schema:
            type: integer
            example: 1
          description: "Page number for pagination (default: 1)."
        - in: query
          name: size
          required: false
          schema:
            type: integer
            example: 20
          description: "Number of results per page (default: 20, max: 100)."
      responses:
        '200':
          description: "All messages ordered from newest to oldest"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    message_id:
                      type: integer
                      example: 1
                    group_id:
                      type: integer
                      example: 1
                    user_id:
                      type: integer
                      example: 101
                    user_message:
                      type: string
                      example: "This is a message"
                    message_time:
                      type: string
                      format: date-time
                      example: "2024-01-15T12:34:56"
        '400':
          description: "Group ID is required"
        '403':
          description: "Unauthorized access. You are not a member of this group."
        '401':
          description: "Unauthorized access or missing token"

    post:
      summary: "Create a new message"
      tags:
        - Messages
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_message:
                  type: string
                  example: "This is a new message"
                group_id:
                  type: integer
                  example: 1
          description: "Group ID and message content."
      responses:
        '201':
          description: "Message created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message created successfully"
                  message_id:
                    type: integer
                    example: 123
        '400':
          description: "Message text or group ID is required"
        '403':
          description: "Unauthorized access. You are not a member of this group."
        '401':
          description: "Unauthorized access or missing token"
  /messages/filter:
    get:
      summary: "Filter messages by date range with pagination"
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: group_id
          required: true
          schema:
            type: integer
            example: 1
          description: "Group ID to filter messages."
        - in: query
          name: start_date
          required: true
          schema:
            type: string
            format: date
            example: "2024-01-01"
          description: "Start date for filtering (YYYY-MM-DD)."
        - in: query
          name: end_date
          required: true
          schema:
            type: string
            format: date
            example: "2024-02-01"
          description: "End date for filtering (YYYY-MM-DD)."
        - in: query
          name: page
          required: false
          schema:
            type: integer
            example: 1
          description: "Page number for pagination (default: 1)."
        - in: query
          name: size
          required: false
          schema:
            type: integer
            example: 20
          description: "Number of results per page (default: 20, max: 100)."
      responses:
        '200':
          description: "Filtered messages"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    message_id:
                      type: integer
                      example: 1
                    user_message:
                      type: string
                      example: "Hello, world!"
                    message_time:
                      type: string
                      format: date-time
                      example: "2024-01-15T12:34:56"
        '400':
          description: "Group ID, start date, and end date are required"
        '403':
          description: "Unauthorized access. You are not a member of this group."
        '401':
          description: "Unauthorized access or missing token"
  /messages/<message_id>:
    delete:
      summary: "Delete a message by its ID"
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: message_id
          required: true
          schema:
            type: integer
          description: "The ID of the message to delete."
      responses:
        '200':
          description: "Message deleted successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Message deleted successfully"
        '403':
          description: "Unauthorized access. Only the creator or a group admin can delete this message."
        '404':
          description: "Message not found"
        '401':
          description: "Unauthorized access or missing token"

    put:
      summary: "Update a message by its ID"
      tags:
        - Messages
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: message_id
          required: true
          schema:
            type: integer
          description: "The ID of the message to update."
      requestBody:
        req
  /tags:
    post:
      summary: "Create a new tag"
      tags:
        - Tags
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag_name:
                  type: string
                  example: "Event Management"
      responses:
        '201':
          description: "Tag created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Tag created successfully"
                  tag_id:
                    type: integer
                    example: 1
        '400':
          description: "Tag name is required"

    get:
      summary: "List all tags with optional search"
      tags:
        - Tags
      parameters:
        - in: query
          name: query
          required: false
          schema:
            type: string
            example: "Event"
          description: "Search string to filter tags by name."
      responses:
        '200':
          description: "A list of tags."
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tag_id:
                      type: integer
                      example: 1
                    tag_name:
                      type: string
                      example: "Event Management"
  /tags/{tag_id}:
    put:
      summary: "Update a tag by its ID"
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tag_id
          required: true
          schema:
            type: integer
          description: "The ID of the tag to update."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag_name:
                  type: string
                  example: "Updated Tag Name"
      responses:
        '200':
          description: "Tag updated successfully"
        '404':
          description: "Tag not found"
        '400':
          description: "Tag name is required"

    delete:
      summary: "Delete a tag by its ID"
      tags:
        - Tags
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: tag_id
          required: true
          schema:
            type: integer
          description: "The ID of the tag to delete."
      responses:
        '200':
          description: "Tag deleted successfully"
        '404':
          description: "Tag not found"